refactor(probabilistic_occupancy_grid_map)!: fix namespace and directory structure (#7872)
* refactor: update include paths for probabilistic occupancy grid map


* refactor: update include paths for probabilistic occupancy grid map


* refactor: update namespace for nodes


* refactor: update namespace for lib


* refactor: remove unused dependency


* refactor: use const pointer for occupancy grid map data


---------
