+++ perception/probabilistic_occupancy_grid_map/lib/costmap_2d/occupancy_grid_map_projective.cpp
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "probabilistic_occupancy_grid_map/pointcloud_based_occupancy_grid_map/occupancy_grid_map_projective.hpp"
+#include "autoware/probabilistic_occupancy_grid_map/costmap_2d/occupancy_grid_map_projective.hpp"
 
-#include "probabilistic_occupancy_grid_map/cost_value.hpp"
-#include "probabilistic_occupancy_grid_map/utils/utils.hpp"
+#include "autoware/probabilistic_occupancy_grid_map/cost_value/cost_value.hpp"
+#include "autoware/probabilistic_occupancy_grid_map/utils/utils.hpp"
 
 #include <autoware/universe_utils/math/unit_conversion.hpp>
 #include <grid_map_costmap_2d/grid_map_costmap_2d.hpp>
@@ -34,6 +34,9 @@
 #endif
 
 #include <algorithm>
+
+namespace autoware::occupancy_grid_map
+{
 namespace costmap_2d
 {
 using sensor_msgs::PointCloud2ConstIterator;
@@ -192,7 +195,7 @@ void OccupancyGridMapProjectiveBlindSpot::updateWithPointCloud(
     }
     raytrace(
       scan_origin.position.x, scan_origin.position.y, ray_end.wx, ray_end.wy,
-      occupancy_cost_value::FREE_SPACE);
+      cost_value::FREE_SPACE);
   }
 
   if (pub_debug_grid_)
@@ -219,7 +222,7 @@ void OccupancyGridMapProjectiveBlindSpot::updateWithPointCloud(
         const auto & source = obstacle_pointcloud_angle_bin.at(dist_index);
         raytrace(
           source.wx, source.wy, source.projected_wx, source.projected_wy,
-          occupancy_cost_value::NO_INFORMATION);
+          cost_value::NO_INFORMATION);
         break;
       }
 
@@ -227,7 +230,7 @@ void OccupancyGridMapProjectiveBlindSpot::updateWithPointCloud(
         const auto & source = obstacle_pointcloud_angle_bin.at(dist_index);
         raytrace(
           source.wx, source.wy, source.projected_wx, source.projected_wy,
-          occupancy_cost_value::NO_INFORMATION);
+          cost_value::NO_INFORMATION);
         continue;
       }
 
@@ -243,13 +246,13 @@ void OccupancyGridMapProjectiveBlindSpot::updateWithPointCloud(
       if (next_raw_distance < next_obstacle_point_distance) {
         const auto & source = obstacle_pointcloud_angle_bin.at(dist_index);
         const auto & target = *raw_distance_iter;
-        raytrace(source.wx, source.wy, target.wx, target.wy, occupancy_cost_value::NO_INFORMATION);
-        setCellValue(target.wx, target.wy, occupancy_cost_value::FREE_SPACE);
+        raytrace(source.wx, source.wy, target.wx, target.wy, cost_value::NO_INFORMATION);
+        setCellValue(target.wx, target.wy, cost_value::FREE_SPACE);
         continue;
       } else {
         const auto & source = obstacle_pointcloud_angle_bin.at(dist_index);
         const auto & target = obstacle_pointcloud_angle_bin.at(dist_index + 1);
-        raytrace(source.wx, source.wy, target.wx, target.wy, occupancy_cost_value::NO_INFORMATION);
+        raytrace(source.wx, source.wy, target.wx, target.wy, cost_value::NO_INFORMATION);
         continue;
       }
     }
@@ -262,7 +265,7 @@ void OccupancyGridMapProjectiveBlindSpot::updateWithPointCloud(
     auto & obstacle_pointcloud_angle_bin = obstacle_pointcloud_angle_bins.at(bin_index);
     for (size_t dist_index = 0; dist_index < obstacle_pointcloud_angle_bin.size(); ++dist_index) {
       const auto & source = obstacle_pointcloud_angle_bin.at(dist_index);
-      setCellValue(source.wx, source.wy, occupancy_cost_value::LETHAL_OBSTACLE);
+      setCellValue(source.wx, source.wy, cost_value::LETHAL_OBSTACLE);
 
       if (dist_index + 1 == obstacle_pointcloud_angle_bin.size()) {
         continue;
@@ -274,7 +277,7 @@ void OccupancyGridMapProjectiveBlindSpot::updateWithPointCloud(
       if (next_obstacle_point_distance <= obstacle_separation_threshold_) {
         const auto & source = obstacle_pointcloud_angle_bin.at(dist_index);
         const auto & target = obstacle_pointcloud_angle_bin.at(dist_index + 1);
-        raytrace(source.wx, source.wy, target.wx, target.wy, occupancy_cost_value::LETHAL_OBSTACLE);
+        raytrace(source.wx, source.wy, target.wx, target.wy, cost_value::LETHAL_OBSTACLE);
         continue;
       }
     }
@@ -299,3 +302,4 @@ void OccupancyGridMapProjectiveBlindSpot::initRosParam(rclcpp::Node & node)
 }
 
 }  // namespace costmap_2d
+}  // namespace autoware::occupancy_grid_map
