+++ perception/probabilistic_occupancy_grid_map/package.xml
@@ -13,9 +13,9 @@
 
   <buildtool_depend>ament_cmake_auto</buildtool_depend>
   <buildtool_depend>autoware_cmake</buildtool_depend>
+  <buildtool_depend>eigen3_cmake_module</buildtool_depend>
 
   <depend>autoware_universe_utils</depend>
-  <depend>eigen3_cmake_module</depend>
   <depend>grid_map_costmap_2d</depend>
   <depend>grid_map_msgs</depend>
   <depend>grid_map_ros</depend>
@@ -27,12 +27,9 @@
   <depend>rclcpp</depend>
   <depend>rclcpp_components</depend>
   <depend>sensor_msgs</depend>
-  <depend>tf2</depend>
   <depend>tf2_eigen</depend>
-  <depend>tf2_geometry_msgs</depend>
   <depend>tf2_ros</depend>
   <depend>tf2_sensor_msgs</depend>
-  <depend>visualization_msgs</depend>
 
   <exec_depend>pointcloud_preprocessor</exec_depend>
   <exec_depend>pointcloud_to_laserscan</exec_depend>
